Implemented grade statistics program, which operates with data from students.txt file, calculates average grade of the group, highest and lowest grades and mode, and these values are written in statistics.txt file after execution.
There was a small problem with this task, because the main task is about calculating the average grade for each student, but there is only one grade for each student present, so I should either write them more grades or leave it as it is. Now I let it be as it is.